Basic importance track work
Hacked together a basic importance track by combining the Ruler and the
BigWig tracks and passing the value from the bigwig file to the part that
makes each sequence. Can be a little buggy especially on Safari, zooming
in and out may fail due to sync issues (?). Need to look more into the numerical
class to see how it handles scrolling/zooming etc. Other TODOs include converting
to heatmap if sufficiently zoomed out, having a scale on the left side and adding
options to change yscale etc.